pyproject.toml
[project]
name = "openai-local-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.60",
    "langchain-ollama>=0.3.3",
    "langchain-text-splitters>=0.3.8",
    "openai-agents>=0.0.15",
]
